+++ db/migrate/20140923144142_add_reviewer_signature_to_agreements.rb
@@ -1,0 +1,5 @@
+class AddReviewerSignatureToAgreements < ActiveRecord::Migration
+  def change
+    add_column :agreements, :reviewer_signature, :text
+  end
+end