+++ db/migrate/20140923151010_add_organization_emails_to_agreements.rb
@@ -1,0 +1,5 @@
+class AddOrganizationEmailsToAgreements < ActiveRecord::Migration
+  def change
+    add_column :agreements, :organization_emails, :text
+  end
+end